收银台 兼容13位计件数量的条码

+++ app/src/main/java/com/easygo/cashier/module/secondary_sreen/UserGoodsScreen.java
@@ -82,7 +82,7 @@ public class UserGoodsScreen extends Presentation {
 
     }
 
-    public void addItem(List<GoodsResponse> result, int weight) {
+    public void addItem(List<GoodsResponse> result, float weight) {
         if(mUserGoodsAdapter != null)
             this.mUserGoodsAdapter.addItem(result, weight);
     }
